requirements.txt
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
databases==0.5.5
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Mako==1.2.0
MarkupSafe==2.1.1
ormar==0.11.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
